requirements.txt
pyyaml
pymongo
hiredis
redis
elasticsearch
happybase
python-etcd
